+++ cmd/mach-composer/terraform.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/labd/mach-composer/generator"
-	"github.com/labd/mach-composer/runner"
+	"github.com/labd/mach-composer/internal/runner"
 	"github.com/spf13/cobra"
 )
