+++ internal/runner/hash.go
